Factor out socks5-client in its own crate
