+++ clients/socks5/Cargo.toml
@@ -12,13 +12,9 @@ path = "src/lib.rs"
 
 [dependencies]
 clap = { version = "4.0", features = ["cargo", "derive"] }
-dirs = "4.0"
-futures = "0.3"
 lazy_static = "1.4.0"
 log = { workspace = true }
-pin-project = "1.0"
 pretty_env_logger = "0.4"
-rand = { version = "0.7.3", features = ["wasm-bindgen"] }
 serde = { workspace = true, features = ["derive"] } # for config serialization/deserialization
 serde_json = { workspace = true }
 tap = "1.0.1"
@@ -31,24 +27,18 @@ nym-bin-common = { path = "../../common/bin-common" }
 client-core = { path = "../../common/client-core", features = ["fs-surb-storage"] }
 nym-coconut-interface = { path = "../../common/coconut-interface" }
 nym-config = { path = "../../common/config" }
-nym-credential-storage = { path = "../../common/credential-storage", optional = true }
 mobile-storage = { path = "../../common/mobile-storage", optional = true }
 nym-credentials = { path = "../../common/credentials" }
 nym-crypto = { path = "../../common/crypto" }
-gateway-client = { path = "../../common/client-libs/gateway-client" }
 gateway-requests = { path = "../../gateway/gateway-requests" }
 nym-network-defaults = { path = "../../common/network-defaults" }
 nym-sphinx = { path = "../../common/nymsphinx" }
 nym-ordered-buffer = { path = "../../common/socks5/ordered-buffer" }
 nym-pemstore = { path = "../../common/pemstore" }
-nym-socks5-proxy-helpers = { path = "../../common/socks5/proxy-helpers" }
-nym-service-providers-common = { path = "../../service-providers/common" }
-nym-socks5-requests = { path = "../../common/socks5/requests" }
-nym-task = { path = "../../common/task" }
 nym-topology = { path = "../../common/topology" }
-validator-client = { path = "../../common/client-libs/validator-client", features = ["nyxd-client"] }
+socks5-client-core = { path = "../../common/socks5-client-core" }
 
 [features]
-default = ["nym-credential-storage"]
+default = []
 eth = []
-mobile = ["mobile-storage", "gateway-client/mobile"]
+mobile = ["mobile-storage", "socks5-client-core/mobile"]
