+++ clients/socks5/src/lib.rs
@@ -1,6 +1,4 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod client;
 pub mod error;
-pub mod socks;
