+++ common/socks5-client-core/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "socks5-client-core"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dirs = "4.0"
+log = { workspace = true }
+pin-project = "1.0"
+rand = { version = "0.7.3", features = ["wasm-bindgen"] }
+serde = { workspace = true, features = ["derive"] } # for config serialization/deserialization
+thiserror = "1.0.34"
+tap = "1.0.1"
+tokio = { version = "1.24.1", features = ["rt-multi-thread", "net", "signal"] }
+
+client-core = { path = "../client-core", features = ["fs-surb-storage"] }
+futures = "0.3"
+gateway-client = { path = "../client-libs/gateway-client" }
+nym-config = { path = "../config" }
+nym-credential-storage = { path = "../credential-storage", optional = true }
+nym-network-defaults = { path = "../network-defaults" }
+nym-socks5-proxy-helpers = { path = "../socks5/proxy-helpers" }
+nym-service-providers-common = { path = "../../service-providers/common" }
+nym-socks5-requests = { path = "../socks5/requests" }
+nym-sphinx = { path = "../nymsphinx" }
+nym-task = { path = "../task" }
+validator-client = { path = "../client-libs/validator-client", features = ["nyxd-client"] }
+
+[features]
+default = ["nym-credential-storage"]
+mobile = ["gateway-client/mobile"]
