+++ common/socks5-client-core/src/socks/authentication.rs
