+++ common/socks5-client-core/src/socks/types.rs
