+++ nym-connect/mobile/src-tauri/src/state.rs
@@ -7,7 +7,7 @@ use futures::SinkExt;
 use tap::TapFallible;
 use tauri::Manager;
 
-use nym_socks5::client::{
+use socks5_client_core::{
     config::Config as Socks5Config, Socks5ControlMessage, Socks5ControlMessageSender,
 };
 use tokio::time::Instant;
